+++ src/javascript/main.js
@@ -12,7 +12,7 @@ License > https://github.com/Graviton-Code-Editor/Graviton-App/blob/master/LICEN
 "use strict"
 
 const GravitonInfo = {
-  date: "191214",
+  date: "191215",
   version: "1.2.0",
   state: "Beta"
 }
@@ -132,13 +132,6 @@ const Settings = require(path.join(
     "windows",
     "market.js"
   )).Market,
-  NewProject = require(path.join(
-    __dirname,
-    "src",
-    "javascript",
-    "windows",
-    "new_project.js"
-  )).NewProject,
   Plugins = require(path.join(
     __dirname,
     "src",
@@ -146,8 +139,7 @@ const Settings = require(path.join(
     "api",
     "plugins.js"
   )).Plugins,
-  Menus = require(path.join(__dirname, "src", "javascript", "api", "menus.js"))
-    .Menus,
+  Menus = require(path.join(__dirname, "src", "javascript", "api", "menus.js")).Menus,
   Explorer = require(path.join(
     __dirname,
     "src",